Remove empty .spec.names.shortNames default from backup CRD.
When keeping a YAML manifest in sync using GitOps tools like ArgoCD, these tools report the manifest being out of sync when the live manifest in the cluster differs from the desired manifest in the source Git/Helm repository.

The `PerconaXtraDBBackup` CRD has an explicit empty list for its `.spec.names.shortNames` property. This is actually the default value, and in that case it apparently gets removed from the live manifest by k8s when applied. This causes the manifest to never become synced in tools like ArgoCD.

The fix is to remove the default empty value from the manifest, which will result in the same configuration on the cluster, but this time in sync with the source manifest.