Update Notes.txt to correct secret name
In this PR: https://github.com/percona/percona-helm-charts/pull/194/commits/fdcb6fbf2da6cf5f2ee92c66f15b9566bcbef03a

a change was made to allow the variable "clusterSecretName" to be configured. Additionally, if the clusterSecretName wasn't specified, the suffix "-secret" was added to the fullname of the database.

The NOTES.txt was emitting on the fullname and not the fullname-secrets so the instructions would be incorrect in every case.